pyproject.toml
[project]
name = "lmstudio-local-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lmstudio>=1.4.1",
    "openai>=1.100.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
